Tests/LibWeb/Text/expected/HTML/focus-viewport-fallback.txt

after intercepted navigation focus reset: body
after fragment focus fallback: body
